refactor: remove advanced configuration options
BREAKING CHANGE: The advanced configuration options have been deprecated in favour of the simple format
