+++ .lintstagedrc.json
@@ -1,7 +1,5 @@
 {
-  "linters": {
-    "*.{js,json,md}": ["prettier --write", "git add"],
-    "*.js": ["npm run lint:base --fix", "git add"],
-    ".*{rc, json}": ["jsonlint", "git add"]
-  }
+  "*.{js,json,md}": ["prettier --write", "git add"],
+  "*.js": ["npm run lint:base --fix", "git add"],
+  ".*{rc, json}": ["jsonlint", "git add"]
 }
