+++ package.json
@@ -34,7 +34,6 @@
     "dedent": "^0.7.0",
     "del": "^4.1.1",
     "execa": "^2.0.1",
-    "is-glob": "^4.0.0",
     "listr": "^0.14.2",
     "listr-update-renderer": "^0.5.0",
     "lodash": "^4.17.11",
@@ -42,8 +41,7 @@
     "micromatch": "^3.1.8",
     "path-is-inside": "^1.0.2",
     "please-upgrade-node": "^3.0.2",
-    "stringify-object": "^3.2.2",
-    "yup": "^0.27.0"
+    "stringify-object": "^3.2.2"
   },
   "devDependencies": {
     "babel-core": "^6.26.3",
