+++ test/__mocks__/my-config.json
@@ -1,5 +1,3 @@
 {
-  "linters": {
-    "*": "mytask"
-  }
+  "*": "mytask"
 }
