+++ test/runAll.2.spec.js
@@ -1,11 +1,10 @@
-import { getConfig } from '../src/getConfig'
 import runAll from '../src/runAll'
 
 jest.unmock('execa')
 
 describe('runAll', () => {
   it('should throw when not in a git directory', async () => {
-    const config = getConfig({ renderer: 'update', cwd: '/' })
+    const config = { cwd: '/' }
     await expect(runAll(config)).rejects.toThrowErrorMatchingSnapshot()
   })
 })
